requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.3.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-gis==0.18
geographiclib==1.52
geopy==2.2.0
idna==3.3
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pymongo==4.1.1
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
